generic version of extract data (written for python2)
Some remarks:
Aggregation will be harder, because this will be different for every type of data.
I do have an idea to solve it tough:
1.For traffic delay: instead of making the sum take the maximum (makes more sense)
2.This is also applicable to traffic accidents_incidents_flitsers
3.For weather data I can't come up with a way to aggregate the data yet, because of the textual decription of the weather...

The inputfiles of the kmi and the accidents are very messy: I will take the liberty of preprocessing them first, because making the cleaner generic to this extent would be the work of a madman.